refactor: reply 컨트롤러의 비즈니스 로직을 service로 옮김

+++ src/main/java/springcafe/reply/service/ReplyService.java
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Service;
 import springcafe.article.model.Article;
 import springcafe.reply.model.Reply;
 import springcafe.reply.repository.ReplyDao;
+import springcafe.user.exception.WrongWriterException;
 
 import java.util.List;
 
@@ -17,7 +18,7 @@ public class ReplyService {
     }
 
 
-    public void create(Article article, String content, String writer){
+    public void saveReply(Article article, String content, String writer) {
         Reply reply = new Reply(content, article.getId(), writer);
         replyDao.save(article, content, reply);
     }
@@ -32,8 +33,15 @@ public class ReplyService {
         return replyDao.findByReplyId(replyId);
     }
 
-    public void delete(Long replyId){
+    public Long deleteReply(Long replyId, String userId){
+
+        Reply replyToDelete = replyDao.findByReplyId(replyId);
+        if(replyToDelete == null|| replyToDelete.matchesWriter(userId)){
+            throw new WrongWriterException("권한이 없습니다.");
+        }
         replyDao.delete(replyId);
+
+        return replyToDelete.getArticleId();// 삭제한 댓글의 게시글로 돌아가도록 articleId 반환
     }
 
 }
